chore(cli): break up `import.rs`
